bugfix: ensure messages are destroyed after handler is done with them
esp_hass_task_event_source() and the user-defined handler are running in
different tasks. try to take binary semaphore before destroying the
message in esp_hass_task_event_source().  otherwise, the message might
be destroyed while the handler is processing the message.

the event loop library keeps a copy of the message, but the message has
a pointer to cJSON object. as a result, the cJSON object can be
corrupted (the memory address to cJSON is copied, but not the memory
content it points to is not).